+++ app/build.gradle.kts
@@ -92,7 +92,6 @@ dependencies{
     /*Auth0*/
     implementation(libs.auth0)
 
-
     // Prueba unitaria
     implementation (platform("androidx.compose:compose-bom:2023.08.00"))
     androidTestImplementation(platform("androidx.compose:compose-bom:2023.06.01"))
