+++ gradle/libs.versions.toml
@@ -1,11 +1,9 @@
 [versions]
 agp = "8.6.0"
 androidxJunit = "1.1.3"
-core = "1.6.1"
 coreTestingVersion = "2.2.0"
 espressoCoreVersion = "3.4.0"
 jetbrainsKotlinxCoroutinesTest = "1.9.0"
-
 kotlin = "2.0.0"
 coreKtx = "1.13.1"
 junit = "4.13.2"
@@ -44,7 +42,6 @@ androidx-ui-test-manifest = { group = "androidx.compose.ui", name = "ui-test-man
 androidx-ui-test-junit4 = { group = "androidx.compose.ui", name = "ui-test-junit4" }
 androidx-navigation-compose = { group = "androidx.navigation", name = "navigation-compose", version.ref = "navigationCompose" }
 androidx-foundation-android = { group = "androidx.compose.foundation", name = "foundation-android", version.ref = "foundationAndroid" }
-
 room = { group = "androidx.room", name = "room-ktx", version.ref = "room" }
 room-runtime = { group = "androidx.room", name = "room-runtime", version.ref = "room" }
 room-compiler = { group = "androidx.room", name = "room-compiler", version.ref = "room" }
